requirements.txt
torch>=2.2.0
streamlit>=1.32.0
plotly>=5.20.0
pandas>=2.0.0
numpy>=1.26.0
matplotlib>=3.8.0
seaborn>=0.13.0
optuna>=3.5.0
scikit-learn>=1.4.0
fpdf2>=2.7.0
kaleido>=0.2.1
